apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
httpx>=0.27,<1.0
apscheduler>=3.10,<4.0
pydantic-settings>=2.4,<3.0
